reorganize into sklearn subpackage and work out the kinks
