+++ spreg/__init__.py
@@ -36,3 +36,6 @@ from .skater_reg import *
 from . import _version
 
 __version__ = _version.get_versions()["version"]
+
+""" Import sklearn subpackage """
+from . import sklearn
