+++ spreg/sklearn/__init__.py
@@ -1,0 +1,9 @@
+"""
+Import all sklearn style functionality as subpackage
+"""
+
+from .durbin import DurbinError, DurbinLag
+from .error import Error
+from .formula import from_formula
+from .lag import Lag
+from .metrics import *
